Restructured directories
The directories have been restructured. UI elements now have their own
sub-directory and so does styles.
