+++ main-process/base/Window.ts
