+++ main-process/tsconfig.json.license
