+++ package.json
@@ -107,8 +107,8 @@
         "react:build": "node scripts/build.js",
         "react:test": "node scripts/test.js",
         "postinstall": "electron-builder install-app-deps",
-        "electron:build": "tsc -p electron",
-        "electron:dev": "concurrently \"cross-env BROWSER=none yarn react:start\" \"wait-on http://localhost:3000 && tsc -p electron -w\" \"wait-on http://localhost:3000 && tsc -p electron && electronmon .\"",
+        "electron:build": "tsc -p main-process",
+        "electron:dev": "concurrently \"cross-env BROWSER=none yarn react:start\" \"wait-on http://localhost:3000 && tsc -p main-process -w\" \"wait-on http://localhost:3000 && tsc -p main-process && electronmon .\"",
         "electron:package:mac": "yarn react:build && tsc -p electron && electron-builder -m -c.extraMetadata.main=build/electron/electron.js",
         "electron:package:win": "yarn react:build && tsc -p electron && electron-builder -w -c.extraMetadata.main=build/electron/electron.js",
         "electron:package:linux": "yarn react:build && tsc -p electron && electron-builder -l -c.extraMetadata.main=build/electron/electron.js"
